+++ CHANGELOG.md
@@ -3,5 +3,9 @@
 # omp-terminal Changelog
 
 ## [Unreleased]
+### Changed
+- Moved "oh-my-pi" from terminal toolbar button to the predefined session dropdown (▼ arrow next to + button)
+- Terminal tab is now named "oh-my-pi" and executes `omp` as a shell command at creation time
+
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)